+++ sentry-android-core/src/main/java/io/sentry/android/core/NetworkBreadcrumbsIntegration.java
@@ -18,7 +18,7 @@ import io.sentry.Integration;
 import io.sentry.SentryLevel;
 import io.sentry.SentryOptions;
 import io.sentry.TypeCheckHint;
-import io.sentry.android.core.internal.util.ConnectivityChecker;
+import io.sentry.android.core.internal.util.AndroidConnectionStatusProvider;
 import io.sentry.util.Objects;
 import java.io.Closeable;
 import java.io.IOException;
@@ -71,7 +71,7 @@ public final class NetworkBreadcrumbsIntegration implements Integration, Closeab
 
       networkCallback = new NetworkBreadcrumbsNetworkCallback(hub, buildInfoProvider);
       final boolean registered =
-          ConnectivityChecker.registerNetworkCallback(
+          AndroidConnectionStatusProvider.registerNetworkCallback(
               context, logger, buildInfoProvider, networkCallback);
 
       // The specific error is logged in the ConnectivityChecker method
@@ -88,7 +88,7 @@ public final class NetworkBreadcrumbsIntegration implements Integration, Closeab
   @Override
   public void close() throws IOException {
     if (networkCallback != null) {
-      ConnectivityChecker.unregisterNetworkCallback(
+      AndroidConnectionStatusProvider.unregisterNetworkCallback(
           context, logger, buildInfoProvider, networkCallback);
       logger.log(SentryLevel.DEBUG, "NetworkBreadcrumbsIntegration remove.");
     }
@@ -210,7 +210,7 @@ public final class NetworkBreadcrumbsIntegration implements Integration, Closeab
       this.signalStrength = strength > -100 ? strength : 0;
       this.isVpn = networkCapabilities.hasTransport(NetworkCapabilities.TRANSPORT_VPN);
       String connectionType =
-          ConnectivityChecker.getConnectionType(networkCapabilities, buildInfoProvider);
+          AndroidConnectionStatusProvider.getConnectionType(networkCapabilities, buildInfoProvider);
       this.type = connectionType != null ? connectionType : "";
     }