+++ sentry-android-core/src/test/java/io/sentry/android/core/SendCachedEnvelopeIntegrationTest.kt
@@ -1,16 +1,21 @@
 package io.sentry.android.core
 
+import io.sentry.IConnectionStatusProvider
+import io.sentry.IConnectionStatusProvider.ConnectionStatus
 import io.sentry.IHub
 import io.sentry.ILogger
 import io.sentry.SendCachedEnvelopeFireAndForgetIntegration.SendFireAndForget
 import io.sentry.SendCachedEnvelopeFireAndForgetIntegration.SendFireAndForgetFactory
 import io.sentry.SentryLevel.DEBUG
+import io.sentry.test.ImmediateExecutorService
+import io.sentry.transport.RateLimiter
 import io.sentry.util.LazyEvaluator
 import org.awaitility.kotlin.await
 import org.mockito.kotlin.any
 import org.mockito.kotlin.eq
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.never
+import org.mockito.kotlin.times
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
 import java.util.concurrent.ExecutionException
@@ -32,8 +37,12 @@ class SendCachedEnvelopeIntegrationTest {
             hasStartupCrashMarker: Boolean = false,
             hasSender: Boolean = true,
             delaySend: Long = 0L,
-            taskFails: Boolean = false
+            taskFails: Boolean = false,
+            useImmediateExecutor: Boolean = false
         ): SendCachedEnvelopeIntegration {
+            if (useImmediateExecutor) {
+                options.executorService = ImmediateExecutorService()
+            }
             options.cacheDirPath = cacheDirPath
             options.setLogger(logger)
             options.isDebug = true
@@ -116,5 +125,91 @@ class SendCachedEnvelopeIntegrationTest {
         // then wait until the async send finishes in background
         await.untilFalse(fixture.flag)
         verify(fixture.sender).send()
+    }
+
+    @Test
+    fun `registers for network connection changes`() {
+        val sut = fixture.getSut(hasStartupCrashMarker = false)
+
+        val connectionStatusProvider = mock<IConnectionStatusProvider>()
+        fixture.options.connectionStatusProvider = connectionStatusProvider
+
+        sut.register(fixture.hub, fixture.options)
+        verify(connectionStatusProvider).addConnectionStatusObserver(any())
+    }
+
+    @Test
+    fun `when theres no network connection does nothing`() {
+        val sut = fixture.getSut(hasStartupCrashMarker = false)
+
+        val connectionStatusProvider = mock<IConnectionStatusProvider>()
+        fixture.options.connectionStatusProvider = connectionStatusProvider
+
+        whenever(connectionStatusProvider.connectionStatus).thenReturn(
+            ConnectionStatus.DISCONNECTED
+        )
+
+        sut.register(fixture.hub, fixture.options)
+        verify(fixture.sender, never()).send()
+    }
+
+    @Test
+    fun `when the network is not disconnected the factory is initialized`() {
+        val sut = fixture.getSut(hasStartupCrashMarker = false)
+
+        val connectionStatusProvider = mock<IConnectionStatusProvider>()
+        fixture.options.connectionStatusProvider = connectionStatusProvider
+
+        whenever(connectionStatusProvider.connectionStatus).thenReturn(
+            ConnectionStatus.UNKNOWN
+        )
+
+        sut.register(fixture.hub, fixture.options)
+        verify(fixture.factory).create(any(), any())
+    }
+
+    @Test
+    fun `whenever network connection status changes, retries sending for relevant statuses`() {
+        val sut = fixture.getSut(hasStartupCrashMarker = false, useImmediateExecutor = true)
+
+        val connectionStatusProvider = mock<IConnectionStatusProvider>()
+        fixture.options.connectionStatusProvider = connectionStatusProvider
+        whenever(connectionStatusProvider.connectionStatus).thenReturn(
+            ConnectionStatus.DISCONNECTED
+        )
+        sut.register(fixture.hub, fixture.options)
+
+        // when there's no connection no factory create call should be done
+        verify(fixture.sender, never()).send()
+
+        // but for any other status processing should be triggered
+        // CONNECTED
+        whenever(connectionStatusProvider.connectionStatus).thenReturn(ConnectionStatus.CONNECTED)
+        sut.onConnectionStatusChanged(ConnectionStatus.CONNECTED)
+        verify(fixture.sender).send()
+
+        // UNKNOWN
+        whenever(connectionStatusProvider.connectionStatus).thenReturn(ConnectionStatus.UNKNOWN)
+        sut.onConnectionStatusChanged(ConnectionStatus.UNKNOWN)
+        verify(fixture.sender, times(2)).send()
+
+        // NO_PERMISSION
+        whenever(connectionStatusProvider.connectionStatus).thenReturn(ConnectionStatus.NO_PERMISSION)
+        sut.onConnectionStatusChanged(ConnectionStatus.NO_PERMISSION)
+        verify(fixture.sender, times(3)).send()
+    }
+
+    @Test
+    fun `when rate limiter is active, does not send envelopes`() {
+        val sut = fixture.getSut(hasStartupCrashMarker = false)
+        val rateLimiter = mock<RateLimiter> {
+            whenever(mock.isActiveForCategory(any())).thenReturn(true)
+        }
+        whenever(fixture.hub.rateLimiter).thenReturn(rateLimiter)
+
+        sut.register(fixture.hub, fixture.options)
+
+        // no factory call should be done if there's rate limiting active
+        verify(fixture.sender, never()).send()
     }
 }