+++ sentry/src/main/java/io/sentry/NoOpSentryClient.java
@@ -2,6 +2,7 @@ package io.sentry;
 
 import io.sentry.protocol.SentryId;
 import io.sentry.protocol.SentryTransaction;
+import io.sentry.transport.RateLimiter;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -51,5 +52,10 @@ final class NoOpSentryClient implements ISentryClient {
       @Nullable Hint hint,
       @Nullable ProfilingTraceData profilingTraceData) {
     return SentryId.EMPTY_ID;
+  }
+
+  @Override
+  public @Nullable RateLimiter getRateLimiter() {
+    return null;
   }
 }