+++ sentry/src/main/java/io/sentry/SendCachedEnvelopeFireAndForgetIntegration.java
@@ -2,16 +2,24 @@ package io.sentry;
 
 import static io.sentry.util.IntegrationUtils.addIntegrationToSdkVersion;
 
+import io.sentry.transport.RateLimiter;
 import io.sentry.util.Objects;
+import java.io.Closeable;
 import java.io.File;
+import java.io.IOException;
 import java.util.concurrent.RejectedExecutionException;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-/** Sends cached events over when your App. is starting. */
-public final class SendCachedEnvelopeFireAndForgetIntegration implements Integration {
+/** Sends cached events over when your App is starting or a network connection is present. */
+public final class SendCachedEnvelopeFireAndForgetIntegration
+    implements Integration, IConnectionStatusProvider.IConnectionStatusObserver, Closeable {
 
   private final @NotNull SendFireAndForgetFactory factory;
+  private @Nullable IConnectionStatusProvider connectionStatusProvider;
+  private @Nullable IHub hub;
+  private @Nullable SentryOptions options;
+  private @Nullable SendFireAndForget sender;
 
   public interface SendFireAndForget {
     void send();
@@ -54,11 +62,10 @@ public final class SendCachedEnvelopeFireAndForgetIntegration implements Integra
     this.factory = Objects.requireNonNull(factory, "SendFireAndForgetFactory is required");
   }
 
-  @SuppressWarnings("FutureReturnValueIgnored")
   @Override
-  public final void register(final @NotNull IHub hub, final @NotNull SentryOptions options) {
-    Objects.requireNonNull(hub, "Hub is required");
-    Objects.requireNonNull(options, "SentryOptions is required");
+  public void register(final @NotNull IHub hub, final @NotNull SentryOptions options) {
+    this.hub = Objects.requireNonNull(hub, "Hub is required");
+    this.options = Objects.requireNonNull(options, "SentryOptions is required");
 
     final String cachedDir = options.getCacheDirPath();
     if (!factory.hasValidPath(cachedDir, options.getLogger())) {
@@ -66,7 +73,58 @@ public final class SendCachedEnvelopeFireAndForgetIntegration implements Integra
       return;
     }
 
-    final SendFireAndForget sender = factory.create(hub, options);
+    options
+        .getLogger()
+        .log(SentryLevel.DEBUG, "SendCachedEventFireAndForgetIntegration installed.");
+    addIntegrationToSdkVersion(getClass());
+
+    connectionStatusProvider = options.getConnectionStatusProvider();
+    connectionStatusProvider.addConnectionStatusObserver(this);
+
+    sender = factory.create(hub, options);
+
+    sendCachedEnvelopes(hub, options);
+  }
+
+  @Override
+  public void close() throws IOException {
+    if (connectionStatusProvider != null) {
+      connectionStatusProvider.removeConnectionStatusObserver(this);
+    }
+  }
+
+  @Override
+  public void onConnectionStatusChanged(
+      final @NotNull IConnectionStatusProvider.ConnectionStatus status) {
+    if (hub != null && options != null) {
+      sendCachedEnvelopes(hub, options);
+    }
+  }
+
+  @SuppressWarnings({"FutureReturnValueIgnored", "NullAway"})
+  private synchronized void sendCachedEnvelopes(
+      final @NotNull IHub hub, final @NotNull SentryOptions options) {
+
+    // skip run only if we're certainly disconnected
+    if (connectionStatusProvider != null
+        && connectionStatusProvider.getConnectionStatus()
+            == IConnectionStatusProvider.ConnectionStatus.DISCONNECTED) {
+      options
+          .getLogger()
+          .log(SentryLevel.INFO, "SendCachedEnvelopeFireAndForgetIntegration, no connection.");
+      return;
+    }
+
+    // in case there's rate limiting active, skip processing
+    final @Nullable RateLimiter rateLimiter = hub.getRateLimiter();
+    if (rateLimiter != null && rateLimiter.isActiveForCategory(DataCategory.All)) {
+      options
+          .getLogger()
+          .log(
+              SentryLevel.INFO,
+              "SendCachedEnvelopeFireAndForgetIntegration, rate limiting active.");
+      return;
+    }
 
     if (sender == null) {
       options.getLogger().log(SentryLevel.ERROR, "SendFireAndForget factory is null.");
@@ -86,11 +144,6 @@ public final class SendCachedEnvelopeFireAndForgetIntegration implements Integra
                       .log(SentryLevel.ERROR, "Failed trying to send cached events.", e);
                 }
               });
-
-      options
-          .getLogger()
-          .log(SentryLevel.DEBUG, "SendCachedEventFireAndForgetIntegration installed.");
-      addIntegrationToSdkVersion(getClass());
     } catch (RejectedExecutionException e) {
       options
           .getLogger()